dataminr/requirements.txt
requests==2.31.0
jwt==1.3.1
google-auth==2.6.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.10.0
